New base class GstOMXSrc and new source plugin GstOMXCameraSrc
Implement OMX camera component source and a matching base class for it.
Note that this is the first public draft of this feature, not something
that is really usable and definitely not something to be considered for
merging upstream right now.

The GstOMXSrc base class is just a dummy class implementing nothing in
itself at this stage. Its only purpose to exist is that it's required in
order to GstOmx plugin loader finding GstOMXCameraSource. Maybe some
generic could could be identified and refactored there sometime.

GstOMXCameraSrc implements source that streams video from OMX camera
component. It's been developed and tested with OpenMAX implementation
by Broadcom on Raspberry Pi. Іt's planned to be tested against other OMX
camera component implementations, at least Bellagio. Currently it's
hard-coded to only handle planar YUV format I420. The plan is to support
all the uncompressed formats supported by the hardware and Gstreamer.
Also on Raspberry Pi, it should be possible to tunnel the camera
component to video encoder component and thus support all the compressed
output formats available. There's probably lots of stuff in
GstOMXVideoEnc that could be recycled and refactored into generic shared
code in this department. See the comments in source code for more
information.

This code probably contains lots of memory leaks. Synchornizing for
concurrent is shoddy. Also error handling is sub standard as is. Go
ahead and fix it ;)