Don't try to deallocate buffers of a port if port is tunneled
gst_omx_component_free() was calling gst_omx_port_deallocate_buffers() on
each port of the component's ports without first checking if the port is
tunneled or not. Only buffers for non-tunneled ports should be
deallocated as the system is handling resource allocation and
deallocation for tunneled ports. This would fail if a component has
both tunneled and non-tunneled port, resulting an assertion being thrown
in gst_omx_port_allocate_buffers_unlocked().